fix(trainer): coerce oversampled batch_size to int; surface tinker resume
- StatefulTaskDataLoader coerces batch_size to int in __init__ so a fractional
  rejection_sample.multiplier (e.g. 1.5) can't reach __len__'s floor-div and
  break len(); the unified_trainer call site also wraps int() to document the
  oversample-then-floor intent.
- tinker: print() the resume / fresh-start milestone instead of logger.info
  (swallowed in console) so resumes are visible like the rest of the loop.

+++ rllm/trainer/unified_trainer.py
@@ -261,7 +261,7 @@ class UnifiedTrainer:
         self._total_training_steps: int | None = None
 
         if self.train_dataset is not None:
-            batch_size = 1 if self.async_config.enable else self.rllm_config.data.train_batch_size * self.rllm_config.rejection_sample.multiplier
+            batch_size = 1 if self.async_config.enable else int(self.rllm_config.data.train_batch_size * self.rllm_config.rejection_sample.multiplier)
             self._train_dataloader = StatefulTaskDataLoader(
                 self.train_dataset,
                 batch_size=batch_size,
